+++ src/components/SnapshotImage.tsx
@@ -1,13 +1,16 @@
+import { Icons } from "@storybook/components";
 import { styled } from "@storybook/theming";
+import React, { ComponentProps } from "react";
 
-export const SnapshotImage = styled.div<{ href?: string; target?: string }>(
+import { CaptureImage, CaptureOverlayImage, ComparisonResult, Test } from "../gql/graphql";
+import { Text } from "./Text";
+
+export const Container = styled.div<{ href?: string; target?: string }>(
   ({ theme }) => ({
     position: "relative",
     display: "flex",
-    justifyContent: "center",
-    alignItems: "center",
     background: "transparent",
-    minHeight: 100,
+    overflow: "hidden",
     margin: 2,
 
     img: {
@@ -16,11 +19,20 @@ export const SnapshotImage = styled.div<{ href?: string; target?: string }>(
     },
     "img + img": {
       position: "absolute",
+      opacity: 0.7,
+      pointerEvents: "none",
     },
     div: {
-      color: theme.color.mediumdark,
+      display: "flex",
+      flexDirection: "column",
+      alignItems: "center",
+      width: "100%",
       margin: "30px 15px",
-      textAlign: "center",
+      color: theme.color.mediumdark,
+      p: {
+        maxWidth: 380,
+        textAlign: "center",
+      },
       svg: {
         width: 28,
         height: 28,
@@ -28,15 +40,19 @@ export const SnapshotImage = styled.div<{ href?: string; target?: string }>(
     },
     "& > svg": {
       position: "absolute",
+      left: "calc(50% - 14px)",
+      top: "calc(50% - 14px)",
       width: 28,
       height: 28,
       color: theme.color.lightest,
       opacity: 0,
       transition: "opacity 0.1s ease-in-out",
+      pointerEvents: "none",
     },
   }),
   ({ href }) =>
     href && {
+      display: "inline-flex",
       "&:hover": {
         "& > svg": {
           opacity: 1,
@@ -47,3 +63,68 @@ export const SnapshotImage = styled.div<{ href?: string; target?: string }>(
       },
     }
 );
+
+interface SnapshotImageProps {
+  componentName: Test["story"]["component"]["name"];
+  storyName: Test["story"]["name"];
+  testUrl: Test["webUrl"];
+  comparisonResult: ComparisonResult;
+  captureImage?: Pick<CaptureImage, "imageUrl" | "imageWidth">;
+  diffImage?: Pick<CaptureOverlayImage, "imageUrl" | "imageWidth">;
+  focusImage?: Pick<CaptureOverlayImage, "imageUrl" | "imageWidth">;
+  diffVisible: boolean;
+  focusVisible: boolean;
+}
+
+export const SnapshotImage = ({
+  componentName,
+  storyName,
+  testUrl,
+  comparisonResult,
+  captureImage,
+  diffImage,
+  focusImage,
+  diffVisible,
+  focusVisible,
+  ...props
+}: SnapshotImageProps & ComponentProps<typeof Container>) => {
+  const hasDiff = captureImage && diffImage && comparisonResult === ComparisonResult.Changed;
+  const hasError = comparisonResult === ComparisonResult.CaptureError;
+  const containerProps = hasDiff ? { as: "a" as any, href: testUrl, target: "_blank" } : {};
+
+  return (
+    <Container {...props} {...containerProps}>
+      {captureImage && (
+        <img
+          alt={`Snapshot for the '${storyName}' story of the '${componentName}' component`}
+          src={captureImage.imageUrl}
+        />
+      )}
+      {hasDiff && diffVisible && (
+        <img
+          alt=""
+          src={diffImage.imageUrl}
+          style={{ maxWidth: `${(diffImage.imageWidth / captureImage.imageWidth) * 100}%` }}
+        />
+      )}
+      {hasDiff && focusImage && focusVisible && (
+        <img
+          alt=""
+          src={focusImage.imageUrl}
+          style={{ maxWidth: `${(focusImage.imageWidth / captureImage.imageWidth) * 100}%` }}
+        />
+      )}
+      {hasDiff && <Icons icon="sharealt" />}
+      {hasError && !captureImage && (
+        <div>
+          <Icons icon="photo" />
+          <Text>
+            A snapshot couldn’t be captured. This often occurs when a story has a code error.
+            Confirm that this story successfully renders in your local Storybook and run the build
+            again.
+          </Text>
+        </div>
+      )}
+    </Container>
+  );
+};