+++ src/screens/VisualTests/VisualTests.tsx
@@ -1,4 +1,5 @@
 import { Icons, Loader } from "@storybook/components";
+import { Icon, TooltipNote, WithTooltip } from "@storybook/design-system";
 // eslint-disable-next-line import/no-unresolved
 import { GitInfo } from "chromatic/node";
 import React, { useCallback, useEffect, useState } from "react";
@@ -8,6 +9,7 @@ import { Button } from "../../components/Button";
 import { Container } from "../../components/Container";
 import { FooterMenu } from "../../components/FooterMenu";
 import { Heading } from "../../components/Heading";
+import { IconButton } from "../../components/IconButton";
 import { ProgressIcon } from "../../components/icons/ProgressIcon";
 import { Bar, Col, Row, Section, Sections, Text } from "../../components/layout";
 import { Text as CenterText } from "../../components/Text";
@@ -29,8 +31,6 @@ import { Warnings } from "./Warnings";
 
 const QueryBuild = graphql(/* GraphQL */ `
   query AddonVisualTestsBuild(
-    $hasBuildId: Boolean!
-    $buildId: ID!
     $projectId: ID!
     $branch: String!
     $gitUserEmailHash: String!
@@ -38,10 +38,7 @@ const QueryBuild = graphql(/* GraphQL */ `
     $storyId: String!
     $testStatuses: [TestStatus!]!
   ) {
-    build(id: $buildId) @include(if: $hasBuildId) {
-      ...BuildFields
-    }
-    project(id: $projectId) @skip(if: $hasBuildId) {
+    project(id: $projectId) {
       name
       lastBuild(
         branches: [$branch]
@@ -130,11 +127,19 @@ const FragmentStoryTestFields = graphql(/* GraphQL */ `
       captureDiff {
         diffImage {
           imageUrl
+          imageWidth
         }
       }
       headCapture {
         captureImage {
           imageUrl
+          imageWidth
+        }
+      }
+      baseCapture {
+        captureImage {
+          imageUrl
+          imageWidth
         }
       }
       viewport {
@@ -154,6 +159,10 @@ const FragmentStoryTestFields = graphql(/* GraphQL */ `
     }
     story {
       storyId
+      name
+      component {
+        name
+      }
     }
   }
 `);
@@ -196,7 +205,6 @@ interface VisualTestsProps {
   storyId: string;
 }
 
-let last: any;
 export const VisualTests = ({
   isStarting,
   lastDevBuildId,
@@ -213,8 +221,6 @@ export const VisualTests = ({
   >({
     query: QueryBuild,
     variables: {
-      hasBuildId: !!lastDevBuildId,
-      buildId: lastDevBuildId || "",
       projectId,
       storyId,
       testStatuses: Object.keys(statusMap) as any as TestStatus[],
@@ -251,7 +257,7 @@ export const VisualTests = ({
     [reviewTest]
   );
 
-  const build = getFragment(FragmentBuildFields, data?.build || data?.project?.lastBuild);
+  const build = getFragment(FragmentBuildFields, data?.project?.lastBuild);
   const isOutdated = build && build.uncommittedHash !== gitInfo.uncommittedHash;
 
   const buildStatusUpdate =
@@ -260,19 +266,15 @@ export const VisualTests = ({
     testsToStatusUpdate(getFragment(FragmentStatusTestFields, build.testsForStatus.nodes));
 
   useEffect(() => {
-    last = {
-      buildStatusUpdate,
-      string: JSON.stringify(buildStatusUpdate),
-    };
-    if (buildStatusUpdate) {
-      updateBuildStatus(buildStatusUpdate);
-    }
+    if (buildStatusUpdate) updateBuildStatus(buildStatusUpdate);
     // We use the stringified version of buildStatusUpdate to do a deep diff
     // eslint-disable-next-line react-hooks/exhaustive-deps
   }, [JSON.stringify(buildStatusUpdate), updateBuildStatus]);
 
   const [settingsVisible, setSettingsVisible] = useState(false);
   const [warningsVisible, setWarningsVisible] = useState(false);
+  const [baselineImageVisible, setBaselineImageVisible] = useState(false);
+  const toggleBaselineImage = () => setBaselineImageVisible(!baselineImageVisible);
 
   if (!build || error) {
     return (
@@ -368,7 +370,9 @@ export const VisualTests = ({
           {...{ tests, isOutdated, startedAt, isStarting, startDevBuild, isBuildFailed }}
         />
         {!isStarting && tests && tests.length > 0 && (
-          <SnapshotComparison {...{ tests, isAccepting, isOutdated, onAccept }} />
+          <SnapshotComparison
+            {...{ tests, isAccepting, isOutdated, onAccept, baselineImageVisible }}
+          />
         )}
       </Section>
 
@@ -381,9 +385,36 @@ export const VisualTests = ({
       <Section>
         <Bar>
           <Col>
-            <Text style={{ marginLeft: 5 }}>Latest snapshot on {build.branch}</Text>
+            <WithTooltip
+              tooltip={<TooltipNote note="Switch snapshot" />}
+              trigger="hover"
+              hasChrome={false}
+            >
+              <IconButton
+                data-testid="button-toggle-snapshot"
+                onClick={() => toggleBaselineImage()}
+              >
+                <Icon icon="transfer" />
+              </IconButton>
+            </WithTooltip>
+          </Col>
+          <Col style={{ overflow: "hidden", whiteSpace: "nowrap" }}>
+            {baselineImageVisible ? (
+              <Text style={{ marginLeft: 5, width: "100%" }}>
+                <b>Baseline</b> Build {build.number} on {build.branch}
+              </Text>
+            ) : (
+              <Text style={{ marginLeft: 5, width: "100%" }}>
+                <b>Latest</b> Build {build.number} on {build.branch}
+              </Text>
+            )}
           </Col>
           {/* <Col push>
+            <WithTooltip
+              tooltip={<TooltipNote note="Render settings" />}
+              trigger="hover"
+              hasChrome={false}
+            >
             <IconButton
               active={settingsVisible}
               aria-label={`${settingsVisible ? "Hide" : "Show"} render settings`}
@@ -394,19 +425,27 @@ export const VisualTests = ({
             >
               <Icons icon="controls" />
             </IconButton>
+          </WithTooltip>
           </Col>
           <Col>
-            <IconButton
-              active={warningsVisible}
-              aria-label={`${warningsVisible ? "Hide" : "Show"} warnings`}
-              onClick={() => {
-                setWarningsVisible(!warningsVisible);
-                setSettingsVisible(false);
-              }}
-              status="warning"
+            <WithTooltip
+              tooltip={<TooltipNote note="View warnings" />}
+              trigger="hover"
+              hasChrome={false}
             >
-              <Icons icon="alert" />2
-            </IconButton>
+              <IconButton
+                active={warningsVisible}
+                aria-label={`${warningsVisible ? "Hide" : "Show"} warnings`}
+                onClick={() => {
+                  setWarningsVisible(!warningsVisible);
+                  setSettingsVisible(false);
+                }}
+                status="warning"
+              >
+                <Icons icon="alert" />2
+              </IconButton>
+
+            </WithTooltip>
           </Col> */}
           <Col push>
             <FooterMenu setAccessToken={setAccessToken} />